criado o container responsável por exibir a carta selecionada
