+++ src/containers/Game.js
@@ -3,6 +3,7 @@ import { connect } from 'react-redux';
 import { loadData, setCard, start, stop, shuffle } from 'actions/game';
 import { Stage } from 'components/Stage';
 import { Menu } from 'components/Menu';
+import ItemSelected from 'containers/ItemSelected';
 import data from 'services/tarot.json';
 import TweenLite from 'gsap';
 
@@ -41,6 +42,7 @@ class Game extends React.Component {
                 <Menu
                     onStart={this.start}
                     onStop={this.stop}
+                    play={this.props.play}
                 />
                 <Stage
                     order={this.props.order}
@@ -49,71 +51,25 @@ class Game extends React.Component {
                     setCard={this.setCard}
                     selected={this.props.selected}
                 />
-                <ItemSelectedC />
+                <ItemSelected />
             </article>
         );
     }
 }
-
-class ItemSelected extends React.Component {
-    constructor(props) {
-        super(props);
-        this.onClose = this.onClose.bind(this);
-    }
-
-    componentWillReceiveProps(nextProps) {
-        if (!!nextProps.item) {
-            TweenLite.to(this.refs.box, 1, { autoAlpha: 1 });
-        }
-    }
-
-    onClose() {
-        TweenLite.to(this.refs.box, 1, { autoAlpha: 0 });
-        this.props.setCard(false);
-    }
-
-    render() {
-        return (
-            <div className="selected-box" ref="box" onClick={this.onClose}>
-                {!!this.props.item &&
-                    <div className="content">
-                        <div className="close" onClick={this.onClose}>&#10005;</div>
-                        <div className="image">
-                            <img src={this.props.item.image} alt={this.props.item.name}/>
-                        </div>
-                        <div className="description">
-                            <h1 className="title">{this.props.item.name}</h1>
-                            <p>
-                                Lorem ipsum dolor sit amet, consectetur adipisicing elit.
-                                Eius et beatae molestias voluptate repudiandae veniam quo suscipit doloremque,
-                                recusandae nisi dolorem dolores ullam debitis autem placeat dicta,
-                                maiores doloribus commodi!
-                            </p>
-
-                            <p>
-                                Lorem ipsum dolor sit amet, consectetur adipisicing elit.
-                                Eius et beatae molestias voluptate repudiandae veniam quo suscipit doloremque,
-                                recusandae nisi dolorem dolores ullam debitis autem placeat dicta,
-                                maiores doloribus commodi! recusandae nisi dolorem dolores
-                                recusandae nisi dolorem dolores
-                            </p>
-                        </div>
-                    </div>
-                }
-            </div>
-        )
-    }
-}
-
-const ItemSelectedC = connect(({ game }) => ({
-    item: game.cards[game.selected],
- }), { setCard })(ItemSelected);
 
 Game.propTypes = {
     order: PropTypes.array,
     cards: PropTypes.object,
     play: PropTypes.bool,
     loadData: PropTypes.func,
+    shuffle: PropTypes.func,
+    start: PropTypes.func,
+    stop: PropTypes.func,
+    setCard: PropTypes.func,
+    selected: PropTypes.oneOfType([
+        PropTypes.bool,
+        PropTypes.number,
+    ]),
 };
 
 const mapStateToProps = ({ game }) => ({
